Here's the authentication part, it is 98% finished
Here is the authentication part, with connectUser() function, and displayMenu() function which display the menu if the user is connected ( connected == 1 ). I think it is 98% finished because I still have to look at making the first letter of the FirstName and LastName uppercase and to look at names containing spaces.